+++ src/main.ts
@@ -16,7 +16,10 @@
 
 import * as core from '@actions/core';
 import { PredefinedAcl } from '@google-cloud/storage';
+import { errorMessage } from '@google-github-actions/actions-utils';
+
 import { Client } from './client';
+import { parseHeadersInput } from './headers';
 
 async function run(): Promise<void> {
   try {
@@ -24,20 +27,50 @@ async function run(): Promise<void> {
     const destination = core.getInput('destination', { required: true });
     const gzip =
       core.getInput('gzip', { required: false }) === 'false' ? false : true;
+    const resumable =
+      core.getInput('resumable', { required: false }) === 'false'
+        ? false
+        : true;
     const predefinedAclInput = core.getInput('predefinedAcl', {
       required: false,
     });
+    const parent =
+      core.getInput('parent', { required: false }).toLowerCase() === 'false'
+        ? false
+        : true;
+    const glob = core.getInput('glob');
+    const concurrency = Number(core.getInput('concurrency')) || 100;
     const predefinedAcl =
       predefinedAclInput === ''
         ? undefined
         : (predefinedAclInput as PredefinedAcl);
-    const serviceAccountKey = core.getInput('credentials');
-    const client = new Client({ credentials: serviceAccountKey });
+    const headersInput = core.getInput('headers', {
+      required: false,
+    });
+    const metadata =
+      headersInput === '' ? undefined : parseHeadersInput(headersInput);
+    const credentials = core.getInput('credentials');
+
+    // Add warning if using credentials
+    if (credentials) {
+      core.warning(
+        '"credentials" input has been deprecated. ' +
+          'Please switch to using google-github-actions/auth which supports both Workload Identity Federation and JSON Key authentication. ' +
+          'For more details, see https://github.com/google-github-actions/upload-cloud-storage#authorization',
+      );
+    }
+
+    const client = new Client({ credentials: credentials });
     const uploadResponses = await client.upload(
       destination,
       path,
+      glob,
       gzip,
+      resumable,
+      parent,
       predefinedAcl,
+      concurrency,
+      metadata,
     );
 
     core.setOutput(
@@ -46,8 +79,11 @@ async function run(): Promise<void> {
         .map((uploadResponse) => uploadResponse[0].name)
         .toString(),
     );
-  } catch (error) {
-    core.setFailed(error.message);
+  } catch (err) {
+    const msg = errorMessage(err);
+    core.setFailed(
+      `google-github-actions/upload-cloud-storage failed with ${msg}`,
+    );
   }
 }