+++ src/upload-helper.ts
@@ -14,9 +14,21 @@
  * limitations under the License.
  */
 
-import { PredefinedAcl, Storage, UploadResponse } from '@google-cloud/storage';
+import * as os from 'os';
 import * as path from 'path';
-import { getFiles } from './util';
+import * as process from 'process';
+
+import {
+  PredefinedAcl,
+  Storage,
+  UploadOptions,
+  UploadResponse,
+} from '@google-cloud/storage';
+import { Metadata } from './headers';
+import { getDestinationFromPath } from './util';
+import globby from 'globby';
+import * as core from '@actions/core';
+import pMap from 'p-map';
 
 /**
  * Wraps interactions with the the GCS library.
@@ -40,74 +52,96 @@ export class UploadHelper {
    *
    * @param bucketName The name of the bucket.
    * @param filename The file path.
-   * @param destination The destination prefix.
+   * @param gzip Gzip files on upload.
+   * @param resumable Allow resuming uploads.
+   * @param destination The destination in GCS to upload file to.
+   * @param predefinedAcl Predefined ACL config.
    * @returns The UploadResponse which contains the file and metadata.
    */
   async uploadFile(
     bucketName: string,
     filename: string,
     gzip: boolean,
+    resumable: boolean,
     destination?: string,
     predefinedAcl?: PredefinedAcl,
+    metadata?: Metadata,
   ): Promise<UploadResponse> {
-    interface UploadOptions {
-      gzip: boolean;
-      destination?: string;
-      predefinedAcl?: PredefinedAcl;
-    }
     const options: UploadOptions = { gzip, predefinedAcl };
+    const normalizedFilePath = path.posix.normalize(filename);
+    // set destination if defined
     if (destination) {
-      // If obj prefix is set, then extract filename and append to prefix.
-      options.destination = `${destination}/${path.posix.basename(filename)}`;
+      options.destination = destination;
     }
+    if (!process.env.UPLOAD_ACTION_NO_LOG) {
+      core.info(
+        `Uploading file: ${normalizedFilePath} to gs://${bucketName}/${
+          destination ? destination : normalizedFilePath
+        }`,
+      );
+    }
+    if (resumable) {
+      options.resumable = true;
+      options.configPath = path.join(
+        os.tmpdir(),
+        `upload-cloud-storage-${process.hrtime.bigint()}.json`,
+      );
+    }
+    if (metadata) {
+      options.metadata = metadata;
+    }
+
     const uploadedFile = await this.storage
       .bucket(bucketName)
-      .upload(filename, options);
+      .upload(normalizedFilePath, options);
     return uploadedFile;
   }
 
   /**
    * Uploads a specified directory to a GCS bucket. Based on
-   * https://github.com/googleapis/nodejs-storage/blob/master/samples/uploadDirectory.js
    *
    * @param bucketName The name of the bucket.
    * @param directoryPath The path of the directory to upload.
-   * @param objectKeyPrefix Optional Prefix for in the GCS bucket.
-   * @param clearExistingFilesFirst Clean files in the prefix before uploading.
+   * @param glob Glob pattern if any.
+   * @param gzip Gzip files on upload.
+   * @param resumable Allow resuming uploads.
+   * @param parent Flag to enable parent dir in destination path.
+   * @param predefinedAcl Predefined ACL config.
+   * @param concurrency Number of files simultaneously uploaded.
    * @returns The list of UploadResponses which contains the file and metadata.
    */
   async uploadDirectory(
     bucketName: string,
     directoryPath: string,
+    glob = '',
     gzip: boolean,
+    resumable: boolean,
     prefix = '',
+    parent = true,
     predefinedAcl?: PredefinedAcl,
+    concurrency = 100,
+    metadata?: Metadata,
   ): Promise<UploadResponse[]> {
-    const pathDirName = path.posix.dirname(directoryPath);
-    // Get list of files in the directory.
-    const filesList = await getFiles(directoryPath);
-
-    const resp = await Promise.all(
-      filesList.map(async (filePath) => {
-        // Get relative path from directoryPath.
-        let destination = `${path.posix.dirname(
-          path.posix.relative(pathDirName, filePath),
-        )}`;
-        // If prefix is set, prepend.
-        if (prefix) {
-          destination = `${prefix}/${destination}`;
-        }
-
-        const uploadResp = await this.uploadFile(
-          bucketName,
-          filePath,
-          gzip,
-          destination,
-          predefinedAcl,
-        );
-        return uploadResp;
-      }),
-    );
-    return resp;
+    // by default we just use directoryPath with empty glob '', which globby evaluates to directory/**/*
+    const filesList = await globby([path.posix.join(directoryPath, glob)]);
+    const uploader = async (filePath: string): Promise<UploadResponse> => {
+      const destination = await getDestinationFromPath(
+        filePath,
+        directoryPath,
+        parent,
+        prefix,
+      );
+      const uploadResp = await this.uploadFile(
+        bucketName,
+        filePath,
+        gzip,
+        resumable,
+        destination,
+        predefinedAcl,
+        Object.assign({}, metadata),
+      );
+      return uploadResp;
+    };
+    return await pMap(filesList, uploader, { concurrency });
   }
 }