+++ tests/client.test.ts
@@ -14,9 +14,11 @@
  * limitations under the License.
  */
 
-import { expect } from 'chai';
 import 'mocha';
+import { expect } from 'chai';
+
 import * as sinon from 'sinon';
+
 import { Client } from '../src/client';
 import { UploadHelper } from '../src/upload-helper';
 import { FAKE_FILE, FAKE_METADATA } from './constants.test';
@@ -24,54 +26,56 @@ import { FAKE_FILE, FAKE_METADATA } from './constants.test';
 /**
  * Unit Test upload method in Client.
  */
-describe('Unit Test Client', function() {
-  afterEach(function() {
+describe('Unit Test Client', function () {
+  afterEach(function () {
     sinon.restore();
   });
 
-  it('initializes with JSON creds', function() {
+  it('initializes with JSON creds', function () {
     const client = new Client({
       credentials: `{"foo":"bar"}`,
     });
     expect(client.storage.authClient.jsonContent).eql({ foo: 'bar' });
   });
 
-  it('initializes with ADC', function() {
+  it('initializes with ADC', function () {
     const client = new Client();
     expect(client.storage.authClient.jsonContent).eql(null);
   });
 
-  it('calls uploadFile', async function() {
+  it('calls uploadFile', async function () {
     const uploadFileStub = sinon
       .stub(UploadHelper.prototype, 'uploadFile')
       .callsFake(() => {
         return Promise.resolve([FAKE_FILE, FAKE_METADATA]);
       });
     const client = new Client();
-    const path = './tests/testdata/test1.txt';
+    const filePath = './tests/testdata/test1.txt';
     const bucketName = 'foo';
     const prefix = 'test-prefix';
-    await client.upload(`${bucketName}/${prefix}`, path, true);
+    await client.upload(`${bucketName}/${prefix}`, filePath);
     expect(uploadFileStub.calledOnce);
     expect(uploadFileStub.firstCall.args[0]).to.equal(bucketName);
-    expect(uploadFileStub.firstCall.args[1]).to.equal(path);
+    expect(uploadFileStub.firstCall.args[1]).to.equal(filePath);
     expect(uploadFileStub.firstCall.args[2]).to.equal(true);
-    expect(uploadFileStub.firstCall.args[3]).to.equal(prefix);
+    expect(uploadFileStub.firstCall.args[3]).to.equal(true);
+    expect(uploadFileStub.firstCall.args[4]).to.equal(`${prefix}/test1.txt`);
   });
 
-  it('calls uploadDirectory', async function() {
+  it('calls uploadDirectory', async function () {
     sinon.stub(UploadHelper.prototype, 'uploadFile').callsFake(() => {
       return Promise.resolve([FAKE_FILE, FAKE_METADATA]);
     });
     const uploadDirSpy = sinon.spy(UploadHelper.prototype, 'uploadDirectory');
     const client = new Client();
-    const path = './tests/testdata';
+    const filePath = './tests/testdata';
     const bucketName = 'foo';
-    await client.upload(bucketName, path, true);
+    await client.upload(bucketName, filePath);
     expect(uploadDirSpy.calledOnce);
     expect(uploadDirSpy.firstCall.args[0]).to.equal(bucketName);
-    expect(uploadDirSpy.firstCall.args[1]).to.equal(path);
-    expect(uploadDirSpy.firstCall.args[2]).to.equal(true);
-    expect(uploadDirSpy.firstCall.args[3]).to.equal('');
+    expect(uploadDirSpy.firstCall.args[1]).to.equal(filePath);
+    expect(uploadDirSpy.firstCall.args[2]).to.equal('');
+    expect(uploadDirSpy.firstCall.args[3]).to.equal(true);
+    expect(uploadDirSpy.firstCall.args[4]).to.equal(true);
   });
 });