+++ CHANGELOG.md
@@ -1,5 +1,39 @@
 # Changelog
 
+## [0.5.0](https://www.github.com/google-github-actions/upload-cloud-storage/compare/v0.4.0...v0.5.0) (2021-11-16)
+
+
+### Features
+
+* Add `headers` option ([#42](https://www.github.com/google-github-actions/upload-cloud-storage/issues/42)) ([#99](https://www.github.com/google-github-actions/upload-cloud-storage/issues/99)) ([531cbeb](https://www.github.com/google-github-actions/upload-cloud-storage/commit/531cbebb6f6d81d00018daceedd5d4ac33b5750a))
+* add WIF docs, add warning for credentials input ([#145](https://www.github.com/google-github-actions/upload-cloud-storage/issues/145)) ([cb8404f](https://www.github.com/google-github-actions/upload-cloud-storage/commit/cb8404f7f7e1b83eb7d9f5a55eff745fab90be97))
+
+
+### Bug Fixes
+
+* absolute path without parent dir ([#108](https://www.github.com/google-github-actions/upload-cloud-storage/issues/108)) ([205a73d](https://www.github.com/google-github-actions/upload-cloud-storage/commit/205a73d2c3105366722104e84f2ad1da16c2ad63))
+
+## [0.4.0](https://www.github.com/google-github-actions/upload-cloud-storage/compare/v0.3.0...v0.4.0) (2021-08-09)
+
+
+### Features
+
+* Add support for configuring resumable uploads ([#34](https://www.github.com/google-github-actions/upload-cloud-storage/issues/34)) ([#36](https://www.github.com/google-github-actions/upload-cloud-storage/issues/36)) ([6623911](https://www.github.com/google-github-actions/upload-cloud-storage/commit/6623911abca8aa04b7e1b453c5bd4c8544cb0811))
+* Refactor action to support upload at root, glob matching ([#33](https://www.github.com/google-github-actions/upload-cloud-storage/issues/33)) ([a789c2a](https://www.github.com/google-github-actions/upload-cloud-storage/commit/a789c2a53adaa50b68f34fe3bfa2cf7b08585b71))
+
+
+### Bug Fixes
+
+* add docs for  resumable ([#54](https://www.github.com/google-github-actions/upload-cloud-storage/issues/54)) ([3bcc856](https://www.github.com/google-github-actions/upload-cloud-storage/commit/3bcc856858d4ef496631306d8ffe071d6e8d013d))
+* skip setup/teardown without auth ([#55](https://www.github.com/google-github-actions/upload-cloud-storage/issues/55)) ([a518f71](https://www.github.com/google-github-actions/upload-cloud-storage/commit/a518f71a726853dbf7be97b32c48a37d62139b15))
+
+## [0.3.0](https://www.github.com/google-github-actions/upload-cloud-storage/compare/v0.2.1...v0.3.0) (2021-05-26)
+
+
+### Features
+
+* Add `predefinedAcl` option ([#28](https://www.github.com/google-github-actions/upload-cloud-storage/issues/28)) ([#31](https://www.github.com/google-github-actions/upload-cloud-storage/issues/31)) ([a2ee322](https://www.github.com/google-github-actions/upload-cloud-storage/commit/a2ee322f4331b6873c769c25eb7e896f0ebddf99))
+
 ### [0.2.1](https://www.github.com/google-github-actions/upload-cloud-storage/compare/v0.2.0...v0.2.1) (2020-11-14)